requirements.txt
flask
werkzeug
docx2pdf
PyPDF2
reportlab
pillow
docx
